# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/tax_categories:
    get:
      tags:
        - TaxCategories
      summary: List TaxCategories
      description: Get a paginated list of TaxCategories with optional filtering and sorting
      operationId: listTaxCategories
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxCategoryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TaxCategories
      summary: Create TaxCategory
      description: Create a new TaxCategory
      operationId: createTaxCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxCategoryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/tax_filing_periods:
    get:
      tags:
        - TaxFilingPeriods
      summary: List TaxFilingPeriods
      description: Get a paginated list of TaxFilingPeriods with optional filtering and sorting
      operationId: listTaxFilingPeriods
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxFilingPeriodList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TaxFilingPeriods
      summary: Create TaxFilingPeriod
      description: Create a new TaxFilingPeriod
      operationId: createTaxFilingPeriod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxFilingPeriodRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxFilingPeriod'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    TaxCategory:
      type: object
      required:
        - id
        - code
        - name
        - tax_kind
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 150
        tax_kind:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreateTaxCategoryRequest:
      type: object
      required:
        - code
        - name
        - tax_kind
        - status
        - metadata
      properties:
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 150
        tax_kind:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    TaxCategoryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaxCategory'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    TaxFilingPeriod:
      type: object
      required:
        - id
        - period
        - next_sequence
        - output_total
        - input_total
        - withholding_total
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        period:
          type: string
          format: date
        npwp:
          nullable: true
          type: string
          maxLength: 16
        taxpayer_segment:
          nullable: true
          type: string
          maxLength: 3
        next_sequence:
          type: integer
          format: int32
        output_total:
          type: string
        input_total:
          type: string
        withholding_total:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreateTaxFilingPeriodRequest:
      type: object
      required:
        - period
        - next_sequence
        - output_total
        - input_total
        - withholding_total
        - status
        - metadata
      properties:
        period:
          type: string
          format: date
        npwp:
          nullable: true
          type: string
          maxLength: 16
        taxpayer_segment:
          nullable: true
          type: string
          maxLength: 3
        next_sequence:
          type: integer
          format: int32
        output_total:
          type: string
        input_total:
          type: string
        withholding_total:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    TaxFilingPeriodList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaxFilingPeriod'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []